Be explicit about pickle, add delay=False
Event.fire() requires the Pickle serializer, so that class instances can
be recreated in background tasks. However, Celery is moving away from
using Pickle because of security concerns.

Add a new parameter "delay" that defaults to True and uses the current
Pickle serializer, but can be set to False to run synchronously. This
allows the caller to write their own task function that doesn't use the
Pickle serializer.

Add some documentation explaining the situation.